+++ object_detection/PVTv2/configs/pvtv2_b0.yaml
@@ -12,7 +12,7 @@ MODEL:
         MLP_RATIO: [8, 8, 4, 4]
         SR_RATIO: [8, 4, 2, 1]
         QKV_BIAS: True
-    DROP_PATH: 0.1
+    DROP_PATH: 0.0 #0.1
 FPN:
     IN_CHANNELS: [32, 64, 160, 256]
 TRAIN:
