+++ object_detection/PVTv2/pvtv2_det.py
@@ -45,7 +45,7 @@ class PVTv2Det(nn.Layer):
         feats = self.neck(self.backbone(x.tensors))
         rpn_out = self.rpnhead(feats, gt)
 
-        if self.config.ROI.PAT_GT_AS_PRO:
+        if self.training and self.config.ROI.PAT_GT_AS_PRO:
             proposals = []
             for proposal, gt_box in zip(rpn_out[0], gt["gt_boxes"]):
                 proposals.append(paddle.concat([proposal, gt_box]))
