+++ object_detection/PVTv2/run_train.sh
@@ -1,0 +1,7 @@
+CUDA_VISIBLE_DEVICES=0 \
+python main_single_gpu.py \
+-cfg='./configs/pvtv2_b0.yaml' \
+-dataset='coco' \
+-batch_size=2 \
+-data_path='/dataset/coco' \
+-pretrained='./pvtv2_b0_maskrcnn'
