+++ object_detection/PVTv2/run_train_multi.sh
@@ -1,0 +1,7 @@
+CUDA_VISIBLE_DEVICES=4,5,6,7 \
+python main_multi_gpu.py \
+-cfg='./configs/pvtv2_b0.yaml' \
+-dataset='coco' \
+-batch_size=4 \
+-data_path='/dataset/coco' \
+-pretrained='./pvtv2_b0_maskrcnn'
